+++ Calculator/src/main/java/application/view/CalcView.java
@@ -3,6 +3,7 @@ package application.view;
 import application.calculator.OpType;
 import java.io.IOException;
 import java.util.function.Consumer;
+import java.util.function.Function;
 import javafx.application.Application;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
@@ -15,6 +16,8 @@ import javafx.stage.Stage;
 
 /**
  * This class provides a resizeable JavaFx GUI for the calculator MVC.
+ * 
+ * @author zlac318
  */
 
 public class CalcView extends Application implements ViewInterface {
@@ -33,14 +36,14 @@ public class CalcView extends Application implements ViewInterface {
   @FXML
   private TextField totalField;
 
-  // Methods for registering an observer and for
-  // accessing user data in the UI
+  // Methods for registering an observer and for accessing user data in the UI
   // These methods build the Observer/Observable pattern
 
-
   @Override
-  public void addCalculateObserver(Runnable f) {
-    calcButton.setOnAction(event -> f.run());
+  public void addCalculateObserver(Function<String, String> evaluate) {
+    calcButton.setOnAction(
+        event -> totalField.setText(evaluate.apply(inputField.getText()).toString()));
+    inputField.getText();
   }
 
   @Override
@@ -72,14 +75,14 @@ public class CalcView extends Application implements ViewInterface {
     Scene scene = new Scene(page);
     scene.getStylesheets().add(getClass().getResource("application.css").toExternalForm());
     primaryStage.setScene(scene);
-    primaryStage.setTitle("MVC/Observer/fxml");
+    primaryStage.setTitle("Calculator (only postfix notation has been implemented)");
     primaryStage.show();
   }
 
   // DO NOT CHANGE ANYTHING BELOW THIS COMMENT
   /////////////////////////////////////////////////////////////////////////////////
+  
   // Block for creating an instance variable for others to use.
-  //
   // Make it a JavaFX singleton. Instance is set by the javafx "initialize" method
   private static volatile CalcView instance = null;