+++ Changelog.md
@@ -6,6 +6,60 @@ All notable changes to this project will be documented in this file.
 
 - No changes.
 
+## [Release 2.0] - 2023-11-28
+
+### Added
+
+- Reverse Polish ([postfix]) calculator.
+
+### Fixed
+
+- Nothing.
+
+### Changed
+
+- Integrated [postfix] calculator with [GUI] & [CLI] views.
+
+### Security
+
+- No changes.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+## [Release 1.0] - 2023-11-23
+
+### Added
+
+- Interfaces for user interactions with the calculator.
+    - AsciiView - interactions through the command line interface.
+    - CalcView - interactions through the [graphical interface].
+
+### Fixed
+
+- The namespace of View.fxml & application.css for the GUI.
+
+### Changed
+
+- Namespaces for classes that make up the user interfaces.
+
+### Security
+
+- No changes.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
 ## [Release 0.1] - 2023-11-22
 
 ### Added
@@ -59,6 +113,12 @@ All notable changes to this project will be documented in this file.
 
 - Nothing.
 
-[release 0.1]: https://gitlab.cim.rhul.ac.uk/zlac318/CS2800/-/tree/release-0?ref_type=heads
-[release 0.0]: https://gitlab.cim.rhul.ac.uk/zlac318/CS2800/-/tree/24431e707d8b7726f0ab34b87b234235b37b23f0
+[Release 2.0]: https://gitlab.cim.rhul.ac.uk/zlac318/CS2800/-/tree/release-2.0
+[CLI]: https://gitlab.cim.rhul.ac.uk/zlac318/CS2800/-/tree/feature-cli?ref_type=heads
+[GUI]: https://gitlab.cim.rhul.ac.uk/zlac318/CS2800/-/tree/feature-gui?ref_type=heads
+[postfix]: https://gitlab.cim.rhul.ac.uk/zlac318/CS2800/-/tree/feature-postfixCalculator?ref_type=heads
+[release 1.0]: https://gitlab.cim.rhul.ac.uk/zlac318/CS2800/-/tree/release-1.0
+[graphical interface]: https://gitlab.cim.rhul.ac.uk/zlac318/CS2800/-/tree/feature-gui?ref_type=heads
+[release 0.1]: https://gitlab.cim.rhul.ac.uk/zlac318/CS2800/-/tree/release-0.1
+[release 0.0]: https://gitlab.cim.rhul.ac.uk/zlac318/CS2800/-/tree/release-0.0
 [Stack assembly]: https://gitlab.cim.rhul.ac.uk/zlac318/CS2800/-/tree/feature-stackAssembly?ref_type=heads